+++ src/main/java/trendravel/photoravel_be/domain/review/service/ReviewService.java
@@ -4,15 +4,14 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import trendravel.photoravel_be.commom.image.service.ImageService;
+import trendravel.photoravel_be.db.respository.location.LocationRepository;
+import trendravel.photoravel_be.db.respository.review.ReviewRepository;
+import trendravel.photoravel_be.db.respository.spot.SpotRepository;
 import trendravel.photoravel_be.domain.review.dto.request.ReviewRequestDto;
 import trendravel.photoravel_be.domain.review.dto.response.ReviewResponseDto;
 import trendravel.photoravel_be.db.review.Review;
 import trendravel.photoravel_be.db.review.enums.ReviewTypes;
-import trendravel.photoravel_be.db.respository.LocationRepository;
-import trendravel.photoravel_be.db.respository.ReviewRepository;
-import trendravel.photoravel_be.db.respository.SpotRepository;
-import trendravel.photoravel_be.commom.service.ImageService;
-
 import java.util.List;
 import java.util.Optional;
